Refactor[#92]: Project File Forder 관리
- 향후 원활한 유지보수를 위한 폴더 관리를 했습니다.
- 독단으로 쓰이는 변수는 private 선언을 하였습니다.
